requirements-onnx.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
onnxruntime==1.18.1     # CPU
pillow==10.4.0
numpy==1.26.4
pydantic==2.8.2
python-multipart==0.0.6

